# cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for a Go Drunk Backend API, demonstrating DDD principles.",
        "title": "Go Drunk Backend API by DDD",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/organizations": {
            "post": {
                "description": "Create a new organization with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organizations"
                ],
                "summary": "Create a new organization",
                "parameters": [
                    {
                        "description": "Organization creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.CreateOrganizationReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns created organization ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format or validation errors",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/organizations/assign-user": {
            "post": {
                "description": "Assign a user to a specific organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organizations"
                ],
                "summary": "Assign user to organization",
                "parameters": [
                    {
                        "description": "User assignment request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.AssignUserToOrganizationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format or validation errors",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/organizations/parent/{parentId}": {
            "get": {
                "description": "Retrieve all organizations that have the specified parent organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organizations by parent ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parent Organization ID",
                        "name": "parentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns list of organizations",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_application_service_dto.OrganizationResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parent organization ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/organizations/{id}": {
            "get": {
                "description": "Retrieve organization details by organization ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organization by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns organization details",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_application_service_dto.OrganizationResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid organization ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    },
                    "404": {
                        "description": "Organization not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update organization details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organizations"
                ],
                "summary": "Update organization",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Organization update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_controller_dto.UpdateOrganizationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request format or validation errors",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    },
                    "404": {
                        "description": "Organization not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an organization by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organizations"
                ],
                "summary": "Delete organization",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid organization ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    },
                    "404": {
                        "description": "Organization not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
        "/organizations/{organizationId}/users": {
            "get": {
                "description": "Retrieve all users belonging to a specific organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organization users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns list of users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_edynnt_veloras-api_internal_auth_application_service_dto.UserResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid organization ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_edynnt_veloras-api_pkg_response.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_edynnt_veloras-api_internal_auth_application_service_dto.OrganizationResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_edynnt_veloras-api_internal_auth_application_service_dto.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "two_fa_enabled": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_edynnt_veloras-api_internal_auth_controller_dto.AssignUserToOrganizationReq": {
            "type": "object",
            "required": [
                "organization_id",
                "user_id"
            ],
            "properties": {
                "organization_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_edynnt_veloras-api_internal_auth_controller_dto.CreateOrganizationReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_edynnt_veloras-api_internal_auth_controller_dto.UpdateOrganizationReq": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_edynnt_veloras-api_pkg_response.APIError": {
            "type": "object",
            "properties": {
                "err": {},
                "errorFlag": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
...
}